Cleanup/Preparation for Convection
Started refactoring the code a bit to prepare for convective heat transfer. Added radiative heat transfer in the Fourier calculations to make it like a vacuum. Before It was just conducting into the environment, but I want it to be more realistic. I also added some more comments and reordered some stuff and added some more parameters