Added Color Map Version + BugFix
Added a color map version that changes the color on a textures pixels instead of instantiating a bunch of game objects. Also fixed a couple of bugs including one where the h calculated for convective heat transfer was the same for all the cells. Fixed it by casting the vertical/horizontal position as a float when checking for a max value.